+++ e2e-tests/kitchen-sink/prepare.js
@@ -1,4 +1,4 @@
-import { spawnSync } from 'node:child_process';
+import { spawn, spawnSync } from 'node:child_process';
 import { cp, mkdir } from 'node:fs/promises';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
@@ -24,4 +24,33 @@ export async function setupTestProject(pathToStoreFiles) {
     stdio: 'inherit',
     shell: true,
   });
+
+  spawn('pnpm', ['start'], {
+    cwd: newPath,
+    stdio: 'inherit',
+    shell: true,
+  });
+
+  await pingMastraServer();
 }
+
+const pingMastraServer = async () => {
+  let intervalCount = 0;
+  const intervalId = setInterval(async () => {
+    if (intervalCount > 10) {
+      clearInterval(intervalId);
+    }
+
+    fetch('http://localhost:4111/agents')
+      .then(res => {
+        if (res.ok) {
+          clearInterval(intervalId);
+        }
+      })
+      .catch(() => {
+        console.log(`Server sleeping... Attempt: ${intervalCount}`);
+      });
+
+    intervalCount++;
+  }, 100);
+};