+++ packages/cli/src/playground/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @mastra/playground
 
+## 0.0.0-create-mastra-e2e-test-20250613091002
+
 ## 0.0.0-create-mastra-e2e-test-20250613085031
 
 ## 0.0.2