+++ packages/mcp/integration-tests/CHANGELOG.md
@@ -1,3 +1,5 @@
 # mcp-server-integration-tests
 
+## 0.0.0-create-mastra-e2e-test-20250613091002
+
 ## 0.0.0-create-mastra-e2e-test-20250613085031